requirements.txt
selenium==4.0.0
robotframework==3.2.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==5.0.0
robotframework-databaselibrary==1.2.4
robotframework-pythonlibcore==2.2.1
robotframework-requests==0.8.0
robotframework-lsp
xlrd==1.2.0
pandas==1.2.1
ibm-db==3.1.4
requests==2.28.1
PyYAML==6.0
robotframework-sshlibrary
virtualenv
robotframework-pabot
rebot
chromedriver-py